Primary and secondary level reads to save battery
Primary read happens immdiatly. If the new level reading is lower than the last reading, the pump is turned on and a secondary reading is taken and used.

This means if the level has not changed, or its gone up, we can safely use the primary reading. If its gone down, we may have lost air in the tube due to atmospheric pressue/temp change and need to pupm the tube to get an accurate reading.

Less usage of the pump saves a lot of battery power. This way we only use it when we need to.